dashboard/requirements.txt
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
psycopg2-binary==2.9.9